requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
psycopg2==2.9.9
requests==2.31.0
urllib3==2.2.0
PyYAML~=6.0.1